+++ src/search/prefs_handler.py
