+++ src/search/views/channel_selection_view.py
