+++ src/search/views/global_search_view.py
