+++ src/search/views/persistent_channel_search_view.py
