+++ src/search/views/results_view.py
@@ -2,7 +2,7 @@ import discord
 import math
 from typing import TYPE_CHECKING
 
-from search.models.qo.thread_search import ThreadSearchQuery
+from search.qo.thread_search import ThreadSearchQuery
 from .components.page_jump_modal import PageJumpModal
 
 if TYPE_CHECKING:
