+++ src/shared/models/thread.py
