+++ src/shared/models/user_search_preferences.py
