+++ src/tag_system/views/components/vote_button.py
