+++ src/tag_system/views/vote_view.py
